@@ -1,7 +1,7 @@
 import React from 'react';
-
 import './Banner.css';
 import Arrow from '../../assets/Arrow'
+
 function Banner() {
   return (
     <div className="bannerParentDiv">
@@ -9,7 +9,7 @@ function Banner() {
         <div className="menuBar">
           <div className="categoryMenu">
             <span>ALL CATEGORIES</span>
-            <Arrow></Arrow> 
+            <Arrow></Arrow>
           </div>
           <div className="otherQuickOptions">
             <span>Cars</span>
@@ -28,7 +28,7 @@ function Banner() {
           />
         </div>
       </div>
-      
+
     </div>
   );
 }
