@@ -1,39 +1,52 @@
 import React, { Fragment, useContext, useState } from 'react';
-import { useNavigate } from 'react-router-dom'
 import './Create.css';
 import Header from '../Header/Header';
-import { FirebaseContext, AuthContext } from '../../Store/ContextFiles'
+import { useNavigate } from 'react-router-dom'
+import { AuthContext } from '../../Store/ContextFiles'
+import { addDoc, collection, getFirestore } from 'firebase/firestore';
+import { getStorage, getDownloadURL, uploadBytes, ref } from 'firebase/storage';
 
 const Create = () => {
-  const [Name, setName] = useState('')
-  const [Category, setCategory] = useState('')
-  const [Price, setPrice] = useState('')
-  const [Image, setImage] = useState('')
+  const [name, setName] = useState('')
+  const [category, setCategory] = useState('')
+  const [price, setPrice] = useState('')
+  const [image, setImage] = useState('')
+  const { User } = useContext(AuthContext)
 
   const navigate = useNavigate()
 
-  const { Firebase } = useContext(FirebaseContext)
-  const { User } = useContext(AuthContext)
   const date = new Date()
-  const db = Firebase.firestore()
+  const db = getFirestore()
+  const storage = getStorage()
 
   const handleSubmit = () => {
-    Firebase.storage().ref(`/Image/${Image.name}`).put(Image).then(({ ref }) => {
-      ref.getDownloadURL().then((url) => {
+    const productCollRef = collection(db, 'products')
+    const path = (`/Image/${image.name}`)
 
-        db.collection('products').add({
-          Name,
-          Category,
-          Price,
-          Url: url,
-          User: User.uid,
-          CreatedAt: date.toDateString()
-        }).then(() => {
-          alert('New Product Added')
-          navigate('/')
+    const storageRef = ref(storage, path)
+
+    uploadBytes(storageRef, image)
+      .then((snapshot) => {
+        const downloadURL = getDownloadURL(snapshot.ref)
+        return downloadURL
+      })
+      .then((downloadURL) => {
+        addDoc(productCollRef, {
+          name,
+          category,
+          price,
+          imgUrl: downloadURL,
+          userId: User.uid,
+          createdAt: date.toDateString()
         })
       })
-    })
+      .then(() => {
+        navigate('/')
+      })
+      .catch((err) => {
+        alert(err.message)
+        console.log(err.message)
+      })
   }
 
   return (
@@ -43,42 +56,38 @@ const Create = () => {
         <div className="centerDiv">
 
           <form>
-            <label htmlFor="fname">Name</label>
-            <br />
+            <label htmlFor="fname">Name</label><br />
             <input
               className="input"
               type="text"
-              value={Name}
+              value={name}
               onChange={(e) => setName(e.target.value)}
               id="fname"
               name="Name" /><br />
 
-            <label htmlFor="fname">Category</label>
-            <br />
+            <label htmlFor="fname">Category</label><br />
             <input
               className="input"
               type="text"
-              value={Category}
+              value={category}
               onChange={(e) => setCategory(e.target.value)}
               id="fname"
               name="category" /><br />
 
-            <label htmlFor="fname">Price</label>
-            <br />
+            <label htmlFor="fname">Price</label><br />
             <input
               className="input"
               type="number"
-              value={Price}
+              value={price}
               onChange={(e) => setPrice(e.target.value)}
               id="fname"
               name="Price" /><br />
           </form>
 
           <br />
-          <img alt="Upload here" src={Image ? URL.createObjectURL(Image) : ''}></img>
-          <br />
-          <input onChange={(e) => setImage(e.target.files[0])} type="file" />
-          <br />
+
+          <img alt="Upload here" src={image ? URL.createObjectURL(image) : ''}></img><br />
+          <input onChange={(e) => setImage(e.target.files[0])} type="file" /><br />
           <button onClick={handleSubmit} className="uploadBtn">upload and Submit</button>
 
         </div>
