@@ -1,39 +1,53 @@
 import './View.css';
-import { PostContext } from '../../Store/productContext';
-import { FirebaseContext } from '../../Store/ContextFiles'
 import React, { useEffect, useState, useContext } from 'react'
+import { PostContext } from '../../Store/productContext';
+import { AuthContext, FirebaseContext } from '../../Store/ContextFiles'
+import { getFirestore, getDocs, collection, where, query } from 'firebase/firestore';
 
 function View() {
-  const { Firebase } = useContext(FirebaseContext)
+  const { app } = useContext(FirebaseContext)
   const { ProdDtls } = useContext(PostContext)
+  const { User } = useContext(AuthContext)
   const [UserDtls, setUserDtls] = useState([])
-  const { User } = ProdDtls
+  const db = getFirestore()
 
+  // Fetching product details from Firestore
   useEffect(() => {
-    Firebase.firestore().collection('user').where('id', '==', User).get().then((res) => {
-      res.forEach(doc => {
-        setUserDtls(doc.data())
+    const q = query(
+      collection(db, 'users'),
+      where('id', '==', User.uid)
+    )
+
+    getDocs(q)
+      .then((docs) => {
+        docs.forEach((doc) => {
+          setUserDtls(doc.data())
+        })
       })
-    })
-  }, [Firebase, User])
+      .catch((err) => {
+        alert(err.message)
+        console.log(err.message)
+      })
+  }, [db, app, User])
 
+  // Rendering
   return (
-
     < div className="viewParentDiv" >
       <div className="imageShowDiv">
-        <img src={ProdDtls.Url} alt="" />
+        <img src={ProdDtls.imgUrl} alt="" />
       </div>
       <div className="rightSection">
         <div className="productDetails">
-          <p>&#x20B9; {ProdDtls.Price} </p>
-          <span>{ProdDtls.Name} </span>
-          <p>{ProdDtls.Category}</p>
+          <p>&#x20B9; {ProdDtls.price} </p>
+          <span>{ProdDtls.name} </span>
+          <p>{ProdDtls.category}</p>
           <span>{ProdDtls.createdAt}</span>
         </div>
         {UserDtls && <div className="contactDetails">
-          <p>{UserDtls.username}</p>
+          <p>{UserDtls.name}</p>
           <p>UID : {UserDtls.id}</p>
           <p>{UserDtls.phone}</p>
+          <p>{UserDtls.email}</p>
         </div>}
       </div>
     </div >
