@@ -1,8 +1,6 @@
 import React from 'react';
-
 import Header from '../Components/Header/Header';
 import Banner from '../Components/Banner/Banner';
-
 import Posts from '../Components/Posts/Posts';
 import Footer from '../Components/Footer/Footer';
 
@@ -14,7 +12,7 @@ function Home(props) {
       <Posts />
       <Footer />
     </div>
-  );
+  )
 }
 
 export default Home;
