@@ -1,10 +1,12 @@
 import { createContext, useState } from "react"
 
+// Firbase Context
 export const FirebaseContext = createContext(null)
+
+// Authentication Context
 export const AuthContext = createContext(null)
 
-
-export default function Context({ children }) {
+export function AuthContextWrapper({ children }) {
     const [User, setUser] = useState(null)
     return (
         <AuthContext.Provider value={{ User, setUser }}>
